third_party/WebKit/LayoutTests/crypto/subtle/ecdsa/generate-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test generating an EC key pair for ECDSA.");

jsTestIsAsync = true;

var algorithmKeyGen = {
    name: "ecdsa",
    namedCurve: "P-256",
};
var extractable = true;

debug("Generating a key pair...");
crypto.subtle.generateKey(algorithmKeyGen, extractable, ["sign", "verify"]).then(function(result) {
    keyPair = result;
    shouldBe("keyPair.toString()", "'[object Object]'");
    shouldBe("keyPair.publicKey.type", "'public'");
    shouldBe("keyPair.publicKey.algorithm.name", "'ECDSA'");
    shouldBe("keyPair.publicKey.algorithm.namedCurve", "'P-256'");
    shouldBe("keyPair.publicKey.usages", '["verify"]');
    shouldBe("keyPair.privateKey.type", "'private'");
    shouldBe("keyPair.privateKey.algorithm.name", "'ECDSA'");
    shouldBe("keyPair.privateKey.algorithm.namedCurve", "'P-256'");
    shouldBe("keyPair.privateKey.usages", '["sign"]');

    debug("\nTesting that custom attributes on keys survive garbage collection...");
    keyPair.publicKey.foo = "bar";
    gc();
    setTimeout(function() {
        gc();
        setTimeout(function() {
            shouldBe("keyPair.publicKey.foo", "'bar'");
            finishJSTest();
        }, 0);
    }, 0);
}).catch(failAndFinishJSTest);

</script>

</body>
</html>
