third_party/WebKit/LayoutTests/crypto/subtle/ecdsa/generate-key-expected.txt

Test generating an EC key pair for ECDSA.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.algorithm.name is 'ECDSA'
PASS keyPair.publicKey.algorithm.namedCurve is 'P-256'
PASS keyPair.publicKey.usages is ["verify"]
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.algorithm.name is 'ECDSA'
PASS keyPair.privateKey.algorithm.namedCurve is 'P-256'
PASS keyPair.privateKey.usages is ["sign"]

Testing that custom attributes on keys survive garbage collection...
PASS keyPair.publicKey.foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
